Changes to config, documentation, and minor code changes to address adam-mesos's comments.
Refactored TaskFactory.java to require both frameworkUser and frameworkSuperUser if nodeManagerUri is set there were too many edge cases to consider and they would likely fail on the slave anyway.  Best to require an explicit config and fail fast with useful error message.  Should address adam-mesos's confusion.

Refactored TaskFactor.java to to the executor from a uri instead of assume it is packaged in the tarball.

Tested but the remote distribution and locally installed both work as expected.