Remove Myriad state after framework shutdown
New Myriad framework fails to start after an earlier one is shutdown.

The Myriad state saves the framework id for the old framework.
When the new framework starts, it tries to re-registers with the old framework id.
This causes an error message similar to

I1029 12:07:08.330493 15884 sched.cpp:819] Got error 'Completed framework
attempted to re-register'
I1029 12:07:08.330507 15884 sched.cpp:1625] Asked to abort the driver
I1029 12:07:08.331507 15884 sched.cpp:861] Aborting framework
'20151029-110656-1986333194-5050-10105-0001'

We need to remove the Myriad state as part of framework shutdown.

Added a method removeMyriadState to MyriadStateStore interface and added
implementation in MyriadFileSystemRMStateStore.