Myriad 162 This corrects how offers are handled when using the framework role parameter.
Prior to this fix if Myriad recieved an offer with resource from the default role and it's declared
role it incorrectly declared all cpu and memory as used by the default role.
This let to TASK_LOST errors, which the logs stated were a result of insufficient resources.

This closes: #32
Review: #32