Myriad 172
This corrects the offer handling in `TaskUtils.getScalarResource`.  Previously, the case were the amount of resources of the frameworkRole was less than the amount used was overlooked.  This caused a positive number to be added to the amount still needed, resulting in more than the desired resource being requested.  Using `Math.max(resourceDoubleValue-used,0)` corrects this behaviour.

Tested and confirmed correct resources were requested.

This closes: #41
Review: https://github.com/apache/incubator-myriad/pull/41